requirements.txt
# Core Dependencies
numpy==1.24.3
pandas==2.0.3
matplotlib==3.7.2
seaborn==0.12.2

# Deep Learning
tensorflow==2.13.0
keras==2.13.1

# Computer Vision
opencv-python==4.8.0.76
Pillow==10.0.0

# Data Processing
scikit-learn==1.3.0
scipy==1.11.2

# Web Framework - Flask
Flask==2.3.3
Werkzeug==2.3.7

# Web Framework - FastAPI
fastapi==0.103.1
uvicorn[standard]==0.23.2
python-multipart==0.0.6

# Utilities
tqdm==4.66.1
python-dotenv==1.0.0

# Jupyter (optional, for development)
jupyter==1.0.0
ipykernel==6.25.2

# Testing (optional)
pytest==7.4.2

# Linting and formatting (optional)
black==23.9.1
flake8==6.1.0
